+++ src/libagh/model.hh
@@ -1,4 +1,4 @@
-// ;-*-C++-*- *  Time-stamp: "2011-01-05 15:57:18 hmmr"
+// ;-*-C++-*- *  Time-stamp: "2011-01-09 12:52:46 hmmr"
 
 /*
  * Author: Andrei Zavada (johnhommer@gmail.com)
@@ -81,7 +81,7 @@ class CSCourse {
 							  req_percent_scored,
 							  swa_laden_pages_before_SWA_0);
 			if ( retval )
-				throw logic_error (string ("layout_measurements returned ") + to_string(retval));
+				throw logic_error (string ("layout_measurements returned ") + to_string((long long int)retval));
 		}
 
 	time_t nth_msmt_start_time( size_t n) const
